requirements.txt
bleach==6.2.0
bokeh==3.8.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
colorcet==3.1.0
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.60.0
greenlet==3.2.4
holoviews==1.21.0
hvplot==0.12.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
linkify-it-py==2.0.3
loguru==0.7.3
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdit-py-plugins==0.5.0
mdurl==0.1.2
narwhals==2.5.0
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
panel==1.8.1
param==2.2.1
pillow==11.3.0
plotly==6.3.0
psycopg2-binary==2.9.10
pyodbc==5.2.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyviz_comms==3.0.6
PyYAML==6.0.2
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
setuptools==78.1.1
six==1.17.0
SQLAlchemy==2.0.43
threadpoolctl==3.6.0
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.5.0
uv==0.8.20
watchdog==6.0.0
webencodings==0.5.1
wheel==0.45.1
win32_setctime==1.2.0
xyzservices==2025.4.0
bleach==6.2.0
bokeh==3.8.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
colorcet==3.1.0
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.60.0
greenlet==3.2.4
holoviews==1.21.0
hvplot==0.12.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
linkify-it-py==2.0.3
loguru==0.7.3
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdit-py-plugins==0.5.0
mdurl==0.1.2
narwhals==2.5.0
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
panel==1.8.1
param==2.2.1
pillow==11.3.0
plotly==6.3.0
psycopg2-binary==2.9.10
pyodbc==5.2.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyviz_comms==3.0.6
PyYAML==6.0.2
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
setuptools==78.1.1
six==1.17.0
SQLAlchemy==2.0.43
threadpoolctl==3.6.0
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.5.0
uv==0.8.20
watchdog==6.0.0
webencodings==0.5.1
wheel==0.45.1
win32_setctime==1.2.0
xyzservices==2025.4.0
